Fix to work with Python 3

+++ test/backend/test_pre.py
@@ -3,6 +3,7 @@ import modloop
 import saliweb.test
 import saliweb.backend
 import py_compile
+import sys
 import os
 import re
 
@@ -20,9 +21,11 @@ class PreProcessTests(saliweb.test.TestCase):
                                "sequence='myseq'", s, re.DOTALL | re.MULTILINE),
                      'Python script does not match regex: ' + s)
         # Make sure that the script contains no syntax errors
-        open('test.py', 'w').write(s)
+        with open('test.py', 'w') as fh:
+            fh.write(s)
         py_compile.compile('test.py', doraise=True)
-        os.unlink('test.pyc')
+        if sys.version_info[0] == 2:
+            os.unlink('test.pyc')
         os.unlink('test.py')
 
     def test_make_sge_script(self):
